vita: update for latest changes

+++ source/skeleton/tween.cpp
@@ -66,7 +66,7 @@ void Tween::play(TweenDirection direction, bool _reset) {
 void Tween::reset() {
 
     if (tween.duration() > 0) {
-        tween.seek(0);
+        tween.seek((float) 0);
     }
 }
 
@@ -125,7 +125,7 @@ void Tween::step() {
 
     auto *object = (C2DObject *) transform;
 
-    auto float4 = tween.step(delta.asMilliseconds(), true);
+    auto float4 = tween.step((float) delta.asMilliseconds(), true);
     if (type == TweenType::Position) {
         transform->setPosition(float4[0], float4[1]);
     } else if (type == TweenType::Rotation) {
